acme-corp-s3-data/strands-app/requirements.txt
strands-sdk>=0.1.0
pandas>=2.0.0
pyarrow>=12.0.0
boto3>=1.26.0
python-dotenv>=1.0.0
streamlit>=1.20.0